Extract color scheming controller concern
